+++ Flow.Launcher.Plugin.ShortcutPlugin/Helper/Constants.cs
@@ -1,4 +1,6 @@
-﻿namespace Flow.Launcher.Plugin.ShortcutPlugin.Utilities;
+﻿using System.Collections.Generic;
+
+namespace Flow.Launcher.Plugin.ShortcutPlugin.Helper;
 
 public static class Constants
 {
@@ -22,4 +24,13 @@ public static class Constants
     public const string ReadmeUrl = GithubRepository + "/blob/master/README.md";
 
     public const string DiscordUsername = "mantelis130";
+
+    public static readonly List<string> IconsExtensions =
+    [
+        ".ico",
+        ".png",
+        ".jpg",
+        ".jpeg",
+        ".webp"
+    ];
 }