+++ Flow.Launcher.Plugin.ShortcutPlugin/Helper/Interfaces/IShortcutHandler.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Flow.Launcher.Plugin.ShortcutPlugin.Models.Shortcuts;
 
-namespace Flow.Launcher.Plugin.ShortcutPlugin.Utilities;
+namespace Flow.Launcher.Plugin.ShortcutPlugin.Helper.Interfaces;
 
 public interface IShortcutHandler
 {