+++ Flow.Launcher.Plugin.ShortcutPlugin/Helper/ShortcutUtilities.cs
@@ -1,7 +1,7 @@
 ﻿using System.Diagnostics;
 using CliWrap;
 
-namespace Flow.Launcher.Plugin.ShortcutPlugin.Utilities;
+namespace Flow.Launcher.Plugin.ShortcutPlugin.Helper;
 
 public static class ShortcutUtilities
 {