+++ Flow.Launcher.Plugin.ShortcutPlugin/Models/Shortcuts/ShortcutType.cs
@@ -5,8 +5,6 @@ public enum ShortcutType
     Directory,
     File,
     Url,
-    Plugin,
     Group,
-    Unspecified,
     Shell
 }