+++ Flow.Launcher.Plugin.ShortcutPlugin/Services/Interfaces/ICommandsService.cs
@@ -5,6 +5,4 @@ namespace Flow.Launcher.Plugin.ShortcutPlugin.Services.Interfaces;
 public interface ICommandsService
 {
     List<Result> ResolveCommand(List<string> arguments, Query query);
-
-    void ReloadPluginData();
 }